Move supporting files into a directory
By putting supporting files into a dirctory, it makes it possible
to run hc as ./hc from the development directory without installing
for development purposes or just to kick the tires.  lhc was chosen
as the directory name because hc was taken by the executable.  Plus
lhc is kind of a funny name for an arbitrary precision calculator.

Signed-off-by: Vernon Mauery <vernon@mauery.com>