Better help parsing and add store/recall
Do some more post tokenizing parsing of the input to better determine
what exactly the user typed.  This makes the next step so the user can
ask for help like 'help s32' because it parses s32 as C_int and returns
help for that.

Some of these changes also allowed us to add store and recall.  So now
there are 52 registers a-zA-Z that are available for use where =@x
stores in register 'x' and @x recalls from register 'x'.

Signed-off-by: Vernon Mauery <vernon@mauery.com>