Allow 'x' arg functions to use a list instead of x items
Sum operates on a list of items that is created by prepare_args using
the first x number of items on the stack.  If x is a List, we can just
use the contents of the list for that function's arguments.  This can be
generalized for any argument that requires 'x' args.

Signed-off-by: Vernon Mauery <vernon@mauery.com>