ivc mode to use '[' and ']' for brackets
Allowing different brackets is a nice touch, but it complicates parsing
when it comes to parsing pasted values.  I am a strong believer that
every value printed by the calculator should be able to be used as input
verbatim.  This means that we would have to either reload the grammar or
allow all bracket values, neither of which sounds good.

Signed-off-by: Vernon Mauery <vernon@mauery.com>